move osinfo to pkg
