+++ exporter/chronicleexporter/go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/goccy/go-json v0.10.5
 	github.com/golang/mock v1.7.0-rc.1
 	github.com/google/uuid v1.6.0
-	github.com/observiq/bindplane-otel-contrib/internal/osinfo v0.0.1
+	github.com/observiq/bindplane-otel-contrib/pkg/osinfo v0.0.1
 	github.com/observiq/bindplane-otel-contrib/pkg/expr v0.0.1
 	github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl v0.147.0
 	github.com/shirou/gopsutil/v3 v3.24.5
@@ -106,6 +106,6 @@ require (
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-replace github.com/observiq/bindplane-otel-contrib/internal/osinfo => ../../internal/osinfo
+replace github.com/observiq/bindplane-otel-contrib/pkg/osinfo => ../../pkg/osinfo
 
 replace github.com/observiq/bindplane-otel-contrib/pkg/expr => ../../pkg/expr
