+++ pkg/osinfo/osinfo.go
