+++ pkg/osinfo/osinfo_test.go
